requirements.txt
fastapi
uvicorn
python-dotenv
notion-client
openai
azure-core
azure-search-documents